+++ src/components/checkbox/Checkbox.js
@@ -1,11 +1,11 @@
-import React, { createRef, PureComponent } from 'react';
-import PropTypes from 'prop-types';
-import theme from './theme.css';
+import { IconCheckmarkMediumOutline, IconCheckmarkSmallOutline, IconMinusSmallOutline } from '@teamleader/ui-icons';
 import cx from 'classnames';
 import omit from 'lodash.omit';
+import PropTypes from 'prop-types';
+import React, { createRef, PureComponent } from 'react';
 import Box, { omitBoxProps, pickBoxProps } from '../box';
-import { TextBody, TextDisplay, TextSmall } from '../typography';
-import { IconCheckmarkSmallOutline, IconCheckmarkMediumOutline, IconMinusSmallOutline } from '@teamleader/ui-icons';
+import { TextBodyCompact, TextDisplay, TextSmall } from '../typography';
+import theme from './theme.css';
 
 class Checkbox extends PureComponent {
   inputNode = createRef();
@@ -36,7 +36,7 @@ class Checkbox extends PureComponent {
 
   render() {
     const { checked, disabled, className, size, label, children, indeterminate, ...others } = this.props;
-    const TextElement = size === 'small' ? TextSmall : size === 'medium' ? TextBody : TextDisplay;
+    const TextElement = size === 'small' ? TextSmall : size === 'medium' ? TextBodyCompact : TextDisplay;
     const IconCheckmark = size === 'large' ? IconCheckmarkMediumOutline : IconCheckmarkSmallOutline;
 
     const restProps = omit(others, ['onChange']);