+++ src/components/label/label.stories.js
@@ -1,9 +1,6 @@
 import React from 'react';
-import { IconInfoBadgedSmallFilled, IconMarkerSmallOutline } from '@teamleader/ui-icons';
-import { Icon, Input, Label, TextSmall, Tooltip } from '../../index';
 import { addStoryInGroup, LOW_LEVEL_BLOCKS } from '../../../.storybook/utils';
-
-const TooltippedIcon = Tooltip(Icon);
+import { Input, Label } from '../../index';
 
 export default {
   component: Label,
@@ -14,16 +11,6 @@ export const DefaultStory = (args) => (
   <Label
     {...args}
     htmlFor="input1"
-    connectedLeft={
-      <Icon>
-        <IconMarkerSmallOutline />
-      </Icon>
-    }
-    connectedRight={
-      <TooltippedIcon tooltip={<TextSmall>This is the label tooltip text</TextSmall>} tooltipSize="small">
-        <IconInfoBadgedSmallFilled />
-      </TooltippedIcon>
-    }
   >
     Input label
     <Input id="input1" placeholder="I am the placeholder" />
@@ -31,5 +18,14 @@ export const DefaultStory = (args) => (
 );
 
 DefaultStory.args = {
-  helpText: 'Optional',
+  tooltip: 'This is the label tooltip text',
 };
+DefaultStory.storyName = 'Label + Input';
+
+export const SoloLabelStory = (args) => (
+  <Label
+    {...args}    
+  >
+    Label
+  </Label>
+);